Fomatting, Hours worked today.
Made it more like a table with printf formatting and added hours worked today to output.

+++ WorkTracker/src/tracker/WorkSession.java
@@ -29,7 +29,7 @@ public class WorkSession {
 		return this.name;
 	}
 
-	public double getHours() {
+	public float getHours() {
 		return this.sessionHours;
 	}
